+++ kubernetes-model-generator/kubernetes-model-certificates/cmd/generate/generate.go
@@ -25,6 +25,7 @@ import (
   apimachineryversion "k8s.io/apimachinery/pkg/version"
 
   certificates "k8s.io/api/certificates/v1"
+  v1alpha1certificates "k8s.io/api/certificates/v1alpha1"
   v1beta1certificates "k8s.io/api/certificates/v1beta1"
 
   "log"
@@ -61,6 +62,8 @@ type Schema struct {
   CertificateSigningRequestStatus             certificates.CertificateSigningRequestStatus
   CertificateSigningRequestCondition          certificates.CertificateSigningRequestCondition
   CertificateSigningRequestList               certificates.CertificateSigningRequestList
+  V1Alpha1ClusterTrustBundle                  v1alpha1certificates.ClusterTrustBundle
+  V1Alpha1ClusterTrustBundleList              v1alpha1certificates.ClusterTrustBundleList
   V1Beta1CertificateSigningRequest            v1beta1certificates.CertificateSigningRequest
   V1Beta1CertificateSigningRequestSpec        v1beta1certificates.CertificateSigningRequestSpec
   V1Beta1CertificateSigningRequestStatus      v1beta1certificates.CertificateSigningRequestStatus
@@ -74,6 +77,7 @@ func main() {
     {"k8s.io/apimachinery/pkg/runtime", "", "io.fabric8.kubernetes.api.model.runtime", "kubernetes_apimachinery_pkg_runtime_", false},
     {"k8s.io/apimachinery/pkg/version", "", "io.fabric8.kubernetes.api.model.version", "kubernetes_apimachinery_pkg_version_", false},
     {"k8s.io/apimachinery/pkg/apis/meta/v1", "", "io.fabric8.kubernetes.api.model", "kubernetes_apimachinery_", false},
+    {"k8s.io/api/certificates/v1alpha1", "certificates.k8s.io", "io.fabric8.kubernetes.api.model.certificates.v1alpha1", "kubernetes_certificates_v1alpha1_", true},
     {"k8s.io/api/certificates/v1beta1", "certificates.k8s.io", "io.fabric8.kubernetes.api.model.certificates.v1beta1", "kubernetes_certificates_v1beta1_", true},
     {"k8s.io/api/certificates/v1", "certificates.k8s.io", "io.fabric8.kubernetes.api.model.certificates.v1", "kubernetes_certificates_v1_", true},
   }